Rename Title component
