+++ src/components/CameraPermissions/Primer/index.js
@@ -1,5 +1,5 @@
 import { h } from 'preact'
-import Title from 'components/Title'
+import PageTitle from 'components/PageTitle'
 import theme from 'components/Theme/style.css'
 import {preventDefaultOnClick} from 'components/utils'
 import Button from 'components/Button'
@@ -9,7 +9,7 @@ import { localised } from '../../../locales'
 
 const Permissions = ({onNext, translate}) => (
   <div className={`${style.container} ${theme.fullHeightContainer}`}>
-    <Title title={translate('webcam_permissions.allow_access')} subTitle={translate('webcam_permissions.enable_webcam_for_selfie')} />
+    <PageTitle title={translate('webcam_permissions.allow_access')} subTitle={translate('webcam_permissions.enable_webcam_for_selfie')} />
     <div className={`${theme.thickWrapper} ${style.bodyWrapper}`}>
       <p className={style.instructions}>{translate('webcam_permissions.click_allow')}</p>
       <div className={style.image}>
