+++ src/components/Complete/index.js
@@ -1,7 +1,7 @@
 import { h, Component } from 'preact'
 
 import { trackComponent } from '../../Tracker'
-import Title from '../Title'
+import PageTitle from '../PageTitle'
 import theme from '../Theme/style.css'
 import style from './style.css'
 import { localised } from '../../locales'
@@ -17,7 +17,7 @@ class Complete extends Component {
     return (
       <div className={style.wrapper}>
           <span className={`${theme.icon}  ${style.icon}`}></span>
-          <Title title={completeMessage} subTitle={completeSubmessage} />
+          <PageTitle title={completeMessage} subTitle={completeSubmessage} />
       </div>
     )
   }
