+++ src/components/Confirm/index.js
@@ -12,7 +12,7 @@ import { poaDocumentTypes } from '../DocumentSelector/documentTypes'
 import Button from '../Button'
 import Error from '../Error'
 import Spinner from '../Spinner'
-import Title from '../Title'
+import PageTitle from '../PageTitle'
 import { trackException, trackComponentAndMode, appendToTracking, sendEvent } from '../../Tracker'
 import { localised } from '../../locales'
 
@@ -62,7 +62,7 @@ const Previews = localised(({capture, retakeAction, confirmAction, error, method
       [style.previewsContainerIsFullScreen]: isFullScreen,
     })}>
       { error.type ? <Error {...{error, withArrow: true}} /> :
-        <Title title={title} subTitle={subTitle} smaller={true} className={style.title}/> }
+        <PageTitle title={title} subTitle={subTitle} smaller={true} className={style.title}/> }
         <CaptureViewer {...{ capture, method, isFullScreen }} />
       <Actions {...{retakeAction, confirmAction, error}} />
     </div>
