+++ src/components/GenericError/index.js
@@ -1,6 +1,6 @@
 import { h, Component } from 'preact'
 
-import Title from '../Title'
+import PageTitle from '../PageTitle'
 import theme from '../Theme/style.css'
 import style from './style.css'
 import errors from '../strings/errors'
@@ -17,7 +17,7 @@ class GenericError extends Component {
     const iconName = icon ? icon : 'genericErrorIcon'
     return (
       <div>
-        <Title title={translate(message)} subTitle={translate(instruction)} />
+        <PageTitle title={translate(message)} subTitle={translate(instruction)} />
         <div className={theme.thickWrapper}>
           <span className={`${theme.icon} ${style[iconName]}`} />
         </div>
