+++ src/components/ProofOfAddress/PoAIntro/index.js
@@ -1,7 +1,7 @@
 import { h } from 'preact'
 import theme from '../../Theme/style.css'
 import style from './style.css'
-import Title from '../../Title'
+import PageTitle from '../../PageTitle'
 import Button from '../../Button'
 import {trackComponent} from '../../../Tracker'
 import {preventDefaultOnClick} from '../../utils'
@@ -9,7 +9,7 @@ import {localised} from '../../../locales'
 
 const PoAIntro = ({country, translate, parseTranslatedTags, nextStep}) => (
   <div className={theme.fullHeightContainer}>
-    <Title title={translate('proof_of_address.intro.title', {
+    <PageTitle title={translate('proof_of_address.intro.title', {
       country: !country || country === 'GBR' ? 'UK' : '',
     })} />
     <div className={style.content}>
