+++ src/components/Uploader/index.js
@@ -10,7 +10,7 @@ import errors from '../strings/errors'
 import { trackComponentAndMode } from '../../Tracker'
 import CustomFileInput from '../CustomFileInput'
 import SwitchDevice from '../crossDevice/SwitchDevice'
-import Title from '../Title'
+import PageTitle from '../PageTitle'
 import Button from '../Button'
 import { getDocumentTypeGroup } from '../DocumentSelector/documentTypes'
 import { localised } from '../../locales'
@@ -98,7 +98,7 @@ class Uploader extends Component {
 
     return (
       <div className={classNames(theme.fullHeightContainer, style.container)}>
-        <Title {...{title, subTitle}}/>
+        <PageTitle {...{title, subTitle}}/>
         <div className={classNames(style.uploaderWrapper, {[style.crossDeviceClient]: !allowCrossDeviceFlow})}>
           { allowCrossDeviceFlow && <SwitchDevice {...{changeFlowTo}}/> }
           <UploadArea
