+++ src/components/Welcome/index.js
@@ -1,5 +1,5 @@
 import { h } from 'preact'
-import Title from '../Title'
+import PageTitle from '../PageTitle'
 import theme from '../Theme/style.css'
 import style from './style.css'
 import {preventDefaultOnClick} from '../utils'
@@ -15,7 +15,7 @@ const Welcome = ({title, descriptions, nextStep, translate}) => {
   const welcomeDescriptions = descriptions ? descriptions : localisedDescriptions(translate)
   return (
     <div>
-      <Title title={welcomeTitle} />
+      <PageTitle title={welcomeTitle} />
       <div className={theme.thickWrapper}>
         <div className={style.text}>
           {welcomeDescriptions.map(description => <p>{description}</p>)}
