+++ src/components/crossDevice/CrossDeviceSubmit/index.js
@@ -2,7 +2,7 @@ import { h, Component } from 'preact'
 import { connect } from 'react-redux'
 import { trackComponent } from '../../../Tracker'
 import {preventDefaultOnClick} from '../../utils'
-import Title from '../../Title'
+import PageTitle from '../../PageTitle'
 import Button from '../../Button'
 import theme from '../../Theme/style.css'
 import style from './style.css'
@@ -35,7 +35,7 @@ class CrossDeviceSubmit extends Component {
     const documentCopy = this.hasMultipleDocuments() ? translate('cross_device.submit.multiple_docs_uploaded') : translate('cross_device.submit.one_doc_uploaded')
     return (
       <div>
-        <Title title={translate('cross_device.submit.title')} subTitle={translate('cross_device.submit.sub_title')} />
+        <PageTitle title={translate('cross_device.submit.title')} subTitle={translate('cross_device.submit.sub_title')} />
         <div className={theme.thickWrapper}>
           <ul className={style.uploadList}>
             <li>
