+++ src/components/crossDevice/MobileConnected/index.js
@@ -2,7 +2,7 @@ import { h } from 'preact'
 
 import theme from '../../Theme/style.css'
 import style from './style.css'
-import Title from '../../Title'
+import PageTitle from '../../PageTitle'
 import { trackComponent } from '../../../Tracker'
 import {preventDefaultOnClick} from '../../utils'
 import { localised } from '../../../locales'
@@ -10,7 +10,7 @@ import { localised } from '../../../locales'
 const MobileConnected = ({translate, back}) => {
   return (
     <div>
-      <Title
+      <PageTitle
         title={translate('cross_device.mobile_connected.title.message')}
         subTitle={translate('cross_device.mobile_connected.title.submessage')}
       />
