+++ src/components/crossDevice/MobileNotificationSent/index.js
@@ -2,14 +2,14 @@ import { h } from 'preact'
 
 import theme from '../../Theme/style.css'
 import style from './style.css'
-import Title from '../../Title'
+import PageTitle from '../../PageTitle'
 import { trackComponent } from '../../../Tracker'
 import {preventDefaultOnClick} from '../../utils'
 import { localised } from '../../../locales'
 
 const MobileNotificationSent = ({sms, translate, previousStep}) =>
   <div>
-    <Title title={translate('cross_device.mobile_notification_sent.title')}/>
+    <PageTitle title={translate('cross_device.mobile_notification_sent.title')}/>
     <div className={theme.thickWrapper}>
       <div className={style.submessage}>{translate('cross_device.mobile_notification_sent.submessage', {number: sms.number})}</div>
       <div className={style.boldMessage}>{translate('cross_device.mobile_notification_sent.bold_message')}</div>
